Move the basemap style into walkers
It was written in wanderers, but there is nothing about it which belongs
to that app, and widnet has been carrying a copy of its ancestor all
along.

`style.rs` becomes a directory, since a style is now three things rather
than one: the spec it is written in, the schema which says what a source
calls things, and the basemap itself. Only the last is opinionated, and
only the last is long.

`Schema` sits beside the basemap rather than inside it - it describes a
source, not this cartography, and someone writing their own style needs it
without wanting our colours. Which is also why its table and its filter
helpers are public now.
